backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-dotenv==1.0.0
supabase==2.0.3
requests==2.31.0
pydantic==2.5.0
gotrue==1.3.1